refactor(#136): extract SMPLXMeshGenerator from mesh_generator.py
Continues the mesh_generator.py decomposition started in PR #172 (which
extracted PrimitiveMeshGenerator).  This PR follows the same pattern for
SMPLXMeshGenerator (~664 LOC), leaving MakeHumanMeshGenerator as the
remaining backend for a future PR.

Pattern (per PR #172):
  - New module: src/.../generators/smplx_mesh_generator.py holds the
    class and its private helpers.
  - mesh_generator.py keeps the shared MeshGeneratorBackend enum,
    GeneratedMeshResult dataclass, MeshGeneratorInterface ABC, and the
    MeshGenerator factory, and now re-exports SMPLXMeshGenerator so
    existing callers keep working.
  - Availability flags (SMPLX_AVAILABLE / TRIMESH_AVAILABLE) and the
    mock-patchable module handles (_smplx_module / _trimesh_module)
    stay on mesh_generator.py; the extracted class dereferences them
    via a module reference so existing
    patch("...mesh_generator._smplx_module") tests keep working.

Stats:
  - mesh_generator.py: 1641 -> 983 LOC
  - smplx_mesh_generator.py: new, 692 LOC
  - 9 new unit tests covering backend_name, re-export identity,
    supported segments, availability gating (smplx missing / model_dir
    missing), failure paths for unavailable smplx/trimesh,
    params=None guard, and a mocked happy path.
  - One existing test in test_security_fixes.py was updated to patch
    the logger on its new home module.
  - Function-size budget baseline entries were moved to the new path
    (no new violations; the entries count was preserved).

Note: PR #172 is still open at time of writing; this PR is based on
main (pre-#172) state.  If #172 lands first the merge should be a
trivial text merge in mesh_generator.py (the two re-export blocks sit
at different file positions).

Remaining backend for next PR: MakeHumanMeshGenerator (~660 LOC).
