+++ scripts/config/function_size_budget_baseline.json
@@ -659,15 +659,15 @@
   "src/shared/python/humanoid_character_builder/core/segment_definitions.py:417:_leg_segments": 72,
   "src/shared/python/humanoid_character_builder/core/segment_definitions.py:549:_arm_joints": 78,
   "src/shared/python/humanoid_character_builder/core/segment_definitions.py:649:_leg_joints": 64,
-  "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:993:_convert_params_to_betas": 54,
-  "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:1088:generate": 80,
-  "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:1218:_segment_smplx_mesh": 96,
-  "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:1360:_fallback_z_segmentation": 95,
   "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:220:_generate_via_api": 103,
   "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:390:_segment_mesh_from_groups": 61,
   "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:493:_segment_by_geometry": 54,
   "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:701:_build_mh_script": 59,
-  "src/shared/python/humanoid_character_builder/generators/mesh_generator.py:883:load_part_segmentation": 60,
+  "src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py:117:load_part_segmentation": 60,
+  "src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py:227:_convert_params_to_betas": 54,
+  "src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py:322:generate": 80,
+  "src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py:452:_segment_smplx_mesh": 96,
+  "src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py:594:_fallback_z_segmentation": 95,
   "src/shared/python/humanoid_character_builder/generators/urdf_generator.py:138:build_model": 57,
   "src/shared/python/humanoid_character_builder/generators/urdf_generator.py:299:_generate_link": 52,
   "src/shared/python/humanoid_character_builder/generators/urdf_generator.py:303:_generate_link": 52,
