+++ src/shared/python/humanoid_character_builder/generators/smplx_mesh_generator.py
@@ -1,0 +1,692 @@
+"""SMPL-X mesh generator for humanoid character builder.
+
+Extracted from ``mesh_generator.py`` as part of issue #136, following
+the pattern established in PR #172 (``PrimitiveMeshGenerator`` extraction).
+
+Note: module-level availability flags (``SMPLX_AVAILABLE``, ``TRIMESH_AVAILABLE``)
+and mock-patchable module handles (``_smplx_module``, ``_trimesh_module``) still
+live on ``mesh_generator`` so existing ``patch("...mesh_generator._smplx_module")``
+tests keep working; we dereference them via the module object at call time.
+"""
+
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+from humanoid_character_builder.core.body_parameters import BodyParameters, GenderModel
+from humanoid_character_builder.generators import mesh_generator as _mg
+from humanoid_character_builder.generators.mesh_generator import (
+    GeneratedMeshResult,
+    MeshGeneratorInterface,
+)
+
+logger = logging.getLogger(__name__)
+
+
+class SMPLXMeshGenerator(MeshGeneratorInterface):
+    """Generate meshes using SMPL-X body model.
+
+    SMPL-X provides a differentiable body model learned from
+    thousands of 3D body scans.
+    """
+
+    # Number of beta shape parameters used
+    NUM_BETAS: int = 10
+
+    # Default SMPL-X mean height (m) and BMI
+    _MEAN_HEIGHT_M: float = 1.75
+    _MEAN_BMI: float = 22.0
+
+    #: Expected vertex count for the standard SMPL-X body model topology.
+    SMPLX_EXPECTED_VERTEX_COUNT: int = 10475
+
+    #: Approximate vertex index ranges for each body segment in the SMPL-X mesh.
+    #: These are estimated based on the 10475-vertex SMPL-X topology.
+    #: Values are (start_inclusive, end_exclusive).
+    SMPLX_SEGMENT_VERTEX_RANGES: dict[str, tuple[int, int]] = {
+        "head": (8000, 10475),
+        "neck": (7500, 8000),
+        "torso": (4000, 7500),
+        "pelvis": (3000, 4000),
+        "left_upper_arm": (1800, 2400),
+        "right_upper_arm": (5300, 5900),
+        "left_forearm": (2400, 2900),
+        "right_forearm": (5900, 6400),
+        "left_hand": (2900, 3500),
+        "right_hand": (6400, 7000),
+        "left_thigh": (100, 600),
+        "right_thigh": (600, 1100),
+        "left_shin": (1100, 1500),
+        "right_shin": (1500, 1800),
+        "left_foot": (0, 100),
+        "right_foot": (50, 150),
+    }
+
+    @classmethod
+    def validate_vertex_ranges(
+        cls,
+        actual_vertex_count: int,
+    ) -> bool:
+        """Validate that hardcoded vertex ranges are consistent with a model.
+
+        Checks that every range in :attr: falls
+        within [0, actual_vertex_count) and that *actual_vertex_count*
+        matches :attr:.
+
+        Args:
+            actual_vertex_count: Total number of vertices in the loaded
+                SMPL-X model mesh.
+
+        Returns:
+            True if all ranges are valid, False otherwise.
+        """
+        assert actual_vertex_count is not None, "actual_vertex_count must be provided"
+        assert actual_vertex_count is not None, "actual_vertex_count must be provided"
+        if actual_vertex_count != cls.SMPLX_EXPECTED_VERTEX_COUNT:
+            logger.warning(
+                "SMPL-X vertex count mismatch: expected %d, got %d. "
+                "Hardcoded segment ranges may be inaccurate.",
+                cls.SMPLX_EXPECTED_VERTEX_COUNT,
+                actual_vertex_count,
+            )
+            return False
+
+        for name, (start, end) in cls.SMPLX_SEGMENT_VERTEX_RANGES.items():
+            if not (0 <= start < actual_vertex_count):
+                logger.warning(
+                    "Segment '%s' start index %d is out of range [0, %d)",
+                    name,
+                    start,
+                    actual_vertex_count,
+                )
+                return False
+            if not (0 < end <= actual_vertex_count):
+                logger.warning(
+                    "Segment '%s' end index %d is out of range (0, %d]",
+                    name,
+                    end,
+                    actual_vertex_count,
+                )
+                return False
+        return True
+
+    @classmethod
+    def load_part_segmentation(
+        cls,
+        model_dir: Path,
+    ) -> dict[str, tuple[int, int]]:
+        """Load official SMPL-X part segmentation from model files if available.
+
+        Looks for smplx_part_segmentation.json or
+        smplx_vert_segmentation.json in *model_dir*.  If found, parses
+        it into the same {segment_name: (start, end)} format used by
+        :attr:.
+
+        If no segmentation file is found, falls back to the hardcoded
+        ranges and emits a warning.
+
+        Args:
+            model_dir: Directory containing SMPL-X model files.
+
+        Returns:
+            Mapping of segment names to (start_inclusive, end_exclusive)
+            vertex index ranges.
+        """
+        assert model_dir is not None, "model_dir must be provided"
+        assert model_dir is not None, "model_dir must be provided"
+        import json
+
+        segmentation_files = [
+            "smplx_part_segmentation.json",
+            "smplx_vert_segmentation.json",
+        ]
+
+        for seg_file in segmentation_files:
+            seg_path = model_dir / seg_file
+            if seg_path.exists():
+                try:
+                    raw = json.loads(seg_path.read_text())
+                    # Expected format: {segment_name: [vertex_indices...]}
+                    ranges: dict[str, tuple[int, int]] = {}
+                    for seg_name, indices in raw.items():
+                        if isinstance(indices, list) and len(indices) > 0:
+                            ranges[seg_name] = (min(indices), max(indices) + 1)
+                    if ranges:
+                        logger.info(
+                            "Loaded SMPL-X part segmentation from %s (%d segments)",
+                            seg_path,
+                            len(ranges),
+                        )
+                        return ranges
+                except (json.JSONDecodeError, OSError, TypeError, ValueError) as exc:
+                    logger.warning(
+                        "Failed to parse segmentation file %s: %s",
+                        seg_path,
+                        exc,
+                    )
+
+        logger.warning(
+            "No SMPL-X part segmentation file found in %s; "
+            "falling back to hardcoded vertex ranges.",
+            model_dir,
+        )
+        return dict(cls.SMPLX_SEGMENT_VERTEX_RANGES)
+
+    @property
+    def backend_name(self) -> str:
+        return "smplx"
+
+    @property
+    def is_available(self) -> bool:
+        """Check availability: smplx package present **and** model dir exists."""
+        if not _mg.SMPLX_AVAILABLE:
+            return False
+        return self.model_dir is None or self.model_dir.exists()
+
+    def __init__(
+        self,
+        model_dir: Path | str | None = None,
+        # Legacy alias kept for backward compat
+        model_path: Path | str | None = None,
+    ) -> None:
+        """Initialize SMPL-X generator.
+
+        Args:
+            model_dir: Directory containing SMPL-X model files (npz format).
+            model_path: Deprecated alias for model_dir.
+        """
+        # Prefer model_dir; fall back to legacy model_path
+        raw = model_dir if model_dir is not None else model_path
+        self.model_dir: Path | None = Path(raw) if raw is not None else None
+
+    # ------------------------------------------------------------------
+    # Static helpers (testable independently of the generate pipeline)
+    # ------------------------------------------------------------------
+
+    @staticmethod
+    def _gender_string(params: BodyParameters) -> str:
+        """Return the SMPL-X gender string for the given body parameters.
+
+        Args:
+            params: Body parameters with gender model.
+
+        Returns:
+            One of "male", "female", or "neutral".
+        """
+        gm = params.gender_model
+        if gm == GenderModel.MALE:
+            return "male"
+        if gm == GenderModel.FEMALE:
+            return "female"
+        return "neutral"
+
+    @staticmethod
+    def _convert_params_to_betas(params: BodyParameters) -> np.ndarray:
+        """Convert BodyParameters to SMPL-X beta shape parameters.
+
+        SMPL-X betas control body shape (NUM_BETAS dimensions):
+        - beta[0]: Height (deviation from mean 1.75 m, scaled by 0.2)
+        - beta[1]: BMI (deviation from mean 22, scaled by 5)
+        - beta[2]: Shoulder width (deviation from 1.0 factor)
+        - beta[3]: Hip width (deviation from 1.0, negative → narrower)
+        - beta[4]: Arm length (deviation from 1.0 factor)
+        - beta[5]: Leg length (deviation from 1.0 factor)
+        - beta[6]: Torso length (deviation from 1.0 factor)
+        - beta[7]: Muscularity (deviation from 0.5 mean)
+        - beta[8]: Body fat factor
+        - beta[9]: Reserved / zero
+
+        Args:
+            params: Body parameters to convert.
+
+        Returns:
+            numpy array of shape (NUM_BETAS,).
+        """
+        betas = np.zeros(SMPLXMeshGenerator.NUM_BETAS)
+
+        # beta[0]: Height
+        mean_h = SMPLXMeshGenerator._MEAN_HEIGHT_M
+        betas[0] = float(np.clip((params.height_m - mean_h) / 0.2, -3.0, 3.0))
+
+        # beta[1]: BMI-based weight
+        bmi = params.mass_kg / (params.height_m**2)
+        mean_bmi = SMPLXMeshGenerator._MEAN_BMI
+        betas[1] = float(np.clip((bmi - mean_bmi) / 5.0, -2.0, 2.0))
+
+        # beta[2]: Shoulder width
+        betas[2] = float(np.clip(params.shoulder_width_factor - 1.0, -1.0, 1.0))
+
+        # beta[3]: Hip width (narrower → negative)
+        betas[3] = float(np.clip(params.hip_width_factor - 1.0, -1.0, 1.0))
+
+        # beta[4]: Arm length
+        betas[4] = float(np.clip(params.arm_length_factor - 1.0, -1.0, 1.0))
+
+        # beta[5]: Leg length
+        betas[5] = float(np.clip(params.leg_length_factor - 1.0, -1.0, 1.0))
+
+        # beta[6]: Torso length
+        betas[6] = float(np.clip(params.torso_length_factor - 1.0, -1.0, 1.0))
+
+        # beta[7]: Muscularity (mean = 0.5)
+        betas[7] = float(np.clip(params.muscularity - 0.5, -0.5, 0.5))
+
+        # beta[8]: Body fat
+        betas[8] = float(np.clip(params.body_fat_factor - 0.3, -0.3, 0.7))
+
+        return betas
+
+    @staticmethod
+    def _segment_mesh(
+        vertices: np.ndarray,
+        faces: np.ndarray,
+        vertex_start: int,
+        vertex_end: int,
+    ) -> tuple[np.ndarray, np.ndarray]:
+        """Extract a segment from a mesh by vertex index range.
+
+        Args:
+            vertices: All mesh vertices, shape (N, 3).
+            faces: All mesh faces, shape (M, 3), 0-indexed vertex refs.
+            vertex_start: Inclusive start of vertex range.
+            vertex_end: Exclusive end of vertex range.
+
+        Returns:
+            (segment_vertices, segment_faces) where segment_faces are
+            re-indexed relative to segment_vertices.
+        """
+        # Find faces where ALL vertices are inside the range
+        assert vertices is not None, "vertices must be provided"
+        assert vertices is not None, "vertices must be provided"
+        in_range = (faces >= vertex_start) & (faces < vertex_end)
+        face_mask = in_range.all(axis=1)
+        seg_faces_global = faces[face_mask]
+
+        if len(seg_faces_global) == 0:
+            return np.zeros((0, 3)), np.zeros((0, 3), dtype=np.int64)
+
+        # Collect unique vertices and re-index
+        unique_verts, inverse = np.unique(seg_faces_global, return_inverse=True)
+        seg_vertices = vertices[unique_verts]
+        seg_faces = inverse.reshape(-1, 3)
+
+        return seg_vertices, seg_faces
+
+    # ------------------------------------------------------------------
+    # Main generate pipeline
+    # ------------------------------------------------------------------
+
+    def generate(
+        self,
+        params: BodyParameters,
+        output_dir: Path,
+        **kwargs: Any,
+    ) -> GeneratedMeshResult:
+        """Generate meshes using SMPL-X body model."""
+        assert params is not None, "params must be provided"
+        assert params is not None, "params must be provided"
+        if not _mg.SMPLX_AVAILABLE:
+            return GeneratedMeshResult(
+                success=False,
+                error_message="smplx package not installed. Install with: pip install smplx",
+            )
+
+        if not _mg.TRIMESH_AVAILABLE:
+            return GeneratedMeshResult(
+                success=False,
+                error_message="trimesh package not installed. Install with: pip install trimesh",
+            )
+
+        output_dir = Path(output_dir)
+        visual_dir = output_dir / "visual"
+        collision_dir = output_dir / "collision"
+        visual_dir.mkdir(parents=True, exist_ok=True)
+        collision_dir.mkdir(parents=True, exist_ok=True)
+
+        try:
+            gender = self._gender_string(params)
+
+            # Use mock-patchable module references
+            model = _mg._smplx_module.create(  # type: ignore[union-attr]
+                str(self.model_dir),
+                model_type="smplx",
+                gender=gender,
+                use_pca=False,
+                flat_hand_mean=True,
+            )
+
+            betas_arr = self._convert_params_to_betas(params)
+            # Prefer torch tensor if torch is available; fall back to numpy
+            try:
+                import torch  # type: ignore[import-untyped]
+
+                betas_input = torch.tensor(betas_arr, dtype=torch.float32).unsqueeze(0)
+            except (ImportError, OSError):
+                # torch not installed or incompatible — pass numpy batch array
+                betas_input = betas_arr[np.newaxis, :].astype(np.float32)
+
+            output = model(betas=betas_input)
+
+            vertices = output.vertices.detach().cpu().numpy().squeeze()  # (N, 3)
+            faces = model.faces  # (M, 3)
+
+            # Scale to target height (SMPL-X is Y-up)
+            current_height = float(vertices[:, 1].max() - vertices[:, 1].min())
+            if current_height > 1e-6:
+                scale_factor = params.height_m / current_height
+                vertices = vertices * scale_factor
+
+            mesh = _mg._trimesh_module.Trimesh(vertices=vertices, faces=faces)  # type: ignore[union-attr]
+
+            return self._segment_smplx_mesh(
+                mesh, model, visual_dir, collision_dir, params
+            )
+
+        except (
+            ValueError,
+            ZeroDivisionError,
+            OverflowError,
+            TypeError,
+            RuntimeError,
+            OSError,
+            ImportError,
+        ) as e:
+            logger.error("SMPL-X generation failed: %s", e)
+            return GeneratedMeshResult(
+                success=False,
+                error_message=f"SMPL-X generation error: {e}",
+            )
+
+    def _find_model_path(self) -> Path | None:
+        """Find SMPL-X model files."""
+        if self.model_dir and self.model_dir.exists():
+            return self.model_dir
+
+        # Common locations
+        search_paths = [
+            Path.home() / ".smplx",
+            Path.home() / "smplx",
+            Path("/usr/share/smplx"),
+            Path("./models/smplx"),
+        ]
+
+        for path in search_paths:
+            if path.exists() and (path / "SMPLX_MALE.npz").exists():
+                self.model_dir = path
+                return path
+
+        return None
+
+    # ------------------------------------------------------------------
+    # SMPL-X joint → segment mapping
+    # ------------------------------------------------------------------
+
+    _SMPLX_JOINT_TO_SEGMENT: dict[int, str] = {
+        0: "pelvis",
+        1: "left_thigh",
+        2: "right_thigh",
+        3: "torso",
+        4: "left_shin",
+        5: "right_shin",
+        6: "torso",
+        7: "left_foot",
+        8: "right_foot",
+        9: "torso",
+        10: "left_foot",
+        11: "right_foot",
+        12: "neck",
+        13: "left_upper_arm",
+        14: "right_upper_arm",
+        15: "head",
+        16: "left_upper_arm",
+        17: "right_upper_arm",
+        18: "left_forearm",
+        19: "right_forearm",
+        20: "left_hand",
+        21: "right_hand",
+    }
+
+    def _segment_smplx_mesh(
+        self,
+        mesh: Any,
+        model: Any,
+        visual_dir: Path,
+        collision_dir: Path,
+        params: BodyParameters,
+    ) -> GeneratedMeshResult:
+        """Segment SMPL-X mesh into body parts using LBS weights or vertex ranges.
+
+        Tries to extract vertex groups from the model's LBS skinning weights.
+        Falls back to the SMPLX_SEGMENT_VERTEX_RANGES approximate ranges.
+        Uses _segment_mesh() + _trimesh_module to build per-segment meshes
+        without calling mesh.submesh(), which is not always available.
+        """
+        assert visual_dir is not None, "visual_dir must be provided"
+        assert visual_dir is not None, "visual_dir must be provided"
+        from humanoid_character_builder.core.segment_definitions import (
+            HUMANOID_SEGMENTS,
+        )
+
+        vertex_groups: dict[str, list[int]] = {}
+
+        # Try to extract groups from LBS skinning weights
+        try:
+            weights = model.lbs_weights.cpu().numpy()
+            vertex_assignments = np.argmax(weights, axis=1)
+            for vertex_idx, joint_idx in enumerate(vertex_assignments):
+                segment_name = self._SMPLX_JOINT_TO_SEGMENT.get(int(joint_idx))
+                if segment_name:
+                    vertex_groups.setdefault(segment_name, []).append(vertex_idx)
+        except (AttributeError, ValueError, ZeroDivisionError, TypeError):
+            # Fallback: use approximate vertex ranges
+            total_verts = len(mesh.vertices) if hasattr(mesh, "vertices") else 10475
+            for seg_name, (start, end) in self.SMPLX_SEGMENT_VERTEX_RANGES.items():
+                # Clamp to actual vertex count
+                end = min(end, total_verts)
+                if end > start:
+                    vertex_groups[seg_name] = list(range(start, end))
+
+        if not vertex_groups:
+            return GeneratedMeshResult(
+                success=False,
+                error_message="SMPL-X segmentation error: no vertex groups produced",
+            )
+
+        # Build per-segment meshes using _segment_mesh + _trimesh_module
+        all_vertices = (
+            np.asarray(mesh.vertices) if hasattr(mesh, "vertices") else np.zeros((0, 3))
+        )
+        all_faces = (
+            np.asarray(mesh.faces)
+            if hasattr(mesh, "faces")
+            else np.zeros((0, 3), dtype=np.int64)
+        )
+
+        mesh_paths: dict[str, Path] = {}
+        collision_paths: dict[str, Path] = {}
+
+        for segment_name, indices in vertex_groups.items():
+            if segment_name not in HUMANOID_SEGMENTS or len(indices) < 10:
+                continue
+            try:
+                seg_verts, seg_faces = self._segment_mesh(
+                    all_vertices, all_faces, min(indices), max(indices) + 1
+                )
+                if len(seg_verts) == 0:
+                    continue
+                submesh = _mg._trimesh_module.Trimesh(  # type: ignore[union-attr]
+                    vertices=seg_verts, faces=seg_faces
+                )
+                vpath = visual_dir / f"{segment_name}.stl"
+                submesh.export(str(vpath))
+                mesh_paths[segment_name] = vpath
+                cpath = collision_dir / f"{segment_name}.stl"
+                submesh.convex_hull.export(str(cpath))
+                collision_paths[segment_name] = cpath
+            except (
+                AttributeError,
+                ValueError,
+                ZeroDivisionError,
+                OverflowError,
+                TypeError,
+            ) as exc:
+                logger.warning("Failed to segment %s: %s", segment_name, exc)
+
+        return GeneratedMeshResult(
+            success=len(mesh_paths) > 0,
+            mesh_paths=mesh_paths,
+            collision_paths=collision_paths,
+            vertex_groups=vertex_groups,
+            metadata={
+                "backend": "smplx",
+                "num_segments": len(mesh_paths),
+            },
+        )
+
+    @staticmethod
+    def _extract_smplx_segments(
+        mesh: Any,
+        visual_dir: Path,
+        collision_dir: Path,
+        vertex_groups: dict[str, list[int]],
+        valid_segments: Any,
+    ) -> tuple[dict[str, Path], dict[str, Path]]:
+        """Extract and export individual segment meshes from SMPL-X vertex groups."""
+        assert visual_dir is not None, "visual_dir must be provided"
+        assert visual_dir is not None, "visual_dir must be provided"
+        mesh_paths: dict[str, Path] = {}
+        collision_paths: dict[str, Path] = {}
+
+        for segment_name, vertices in vertex_groups.items():
+            if segment_name not in valid_segments or len(vertices) < 10:
+                continue
+
+            try:
+                vertex_set = set(vertices)
+                face_mask = [
+                    i
+                    for i, face in enumerate(mesh.faces)
+                    if any(v in vertex_set for v in face)
+                ]
+
+                if not face_mask:
+                    continue
+
+                submesh = mesh.submesh([face_mask], append=True)
+
+                visual_path = visual_dir / f"{segment_name}.stl"
+                submesh.export(str(visual_path))
+                mesh_paths[segment_name] = visual_path
+
+                collision_mesh = submesh.convex_hull
+                collision_path = collision_dir / f"{segment_name}.stl"
+                collision_mesh.export(str(collision_path))
+                collision_paths[segment_name] = collision_path
+
+            except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
+                logger.warning(f"Failed to extract segment {segment_name}: {e}")
+
+        return mesh_paths, collision_paths
+
+    def _fallback_z_segmentation(
+        self,
+        mesh: Any,
+        visual_dir: Path,
+        collision_dir: Path,
+        params: BodyParameters,
+    ) -> GeneratedMeshResult:
+        """Fallback segmentation using z-coordinate slicing."""
+        assert visual_dir is not None, "visual_dir must be provided"
+        assert visual_dir is not None, "visual_dir must be provided"
+        from humanoid_character_builder.core.segment_definitions import (
+            HUMANOID_SEGMENTS,
+        )
+
+        mesh_paths = {}
+        collision_paths = {}
+
+        bounds = mesh.bounds
+        height = bounds[1][1] - bounds[0][1]  # SMPL-X uses Y as up
+
+        # Define segment z-ranges (normalized 0-1 from feet to head)
+        segment_ranges = {
+            "left_foot": (0.0, 0.06),
+            "right_foot": (0.0, 0.06),
+            "left_shin": (0.06, 0.25),
+            "right_shin": (0.06, 0.25),
+            "left_thigh": (0.25, 0.47),
+            "right_thigh": (0.25, 0.47),
+            "pelvis": (0.47, 0.55),
+            "torso": (0.55, 0.80),
+            "neck": (0.80, 0.85),
+            "head": (0.85, 1.0),
+            "left_upper_arm": (0.70, 0.80),
+            "right_upper_arm": (0.70, 0.80),
+            "left_forearm": (0.65, 0.70),
+            "right_forearm": (0.65, 0.70),
+            "left_hand": (0.55, 0.65),
+            "right_hand": (0.55, 0.65),
+        }
+
+        vertices = mesh.vertices
+
+        for segment_name, (y_low, y_high) in segment_ranges.items():
+            if segment_name not in HUMANOID_SEGMENTS:
+                continue
+
+            y_min = bounds[0][1] + y_low * height
+            y_max = bounds[0][1] + y_high * height
+
+            # Find vertices in this range
+            mask = (vertices[:, 1] >= y_min) & (vertices[:, 1] <= y_max)
+
+            # For left/right segments, also filter by x
+            if "left" in segment_name:
+                mask &= vertices[:, 0] > 0
+            elif "right" in segment_name:
+                mask &= vertices[:, 0] < 0
+
+            vertex_indices = list(mask.nonzero()[0])
+
+            if len(vertex_indices) < 10:
+                continue
+
+            try:
+                # Find faces using these vertices
+                vertex_set = set(vertex_indices)
+                face_mask = [
+                    i
+                    for i, face in enumerate(mesh.faces)
+                    if any(v in vertex_set for v in face)
+                ]
+
+                if not face_mask:
+                    continue
+
+                submesh = mesh.submesh([face_mask], append=True)
+
+                visual_path = visual_dir / f"{segment_name}.stl"
+                submesh.export(str(visual_path))
+                mesh_paths[segment_name] = visual_path
+
+                collision_mesh = submesh.convex_hull
+                collision_path = collision_dir / f"{segment_name}.stl"
+                collision_mesh.export(str(collision_path))
+                collision_paths[segment_name] = collision_path
+
+            except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
+                logger.warning(f"Failed z-segmentation for {segment_name}: {e}")
+
+        return GeneratedMeshResult(
+            success=len(mesh_paths) > 0,
+            mesh_paths=mesh_paths,
+            collision_paths=collision_paths,
+            metadata={"backend": "smplx", "method": "z_segmentation"},
+        )
+
+    def get_supported_segments(self) -> list[str]:
+        """Return segment names defined in SMPLX_SEGMENT_VERTEX_RANGES."""
+        return list(self.SMPLX_SEGMENT_VERTEX_RANGES.keys())
